+++ java-oop-ru/interfaces/src/main/java/exercise/Flat.java
@@ -18,8 +18,8 @@ public class Flat implements Home {
     }
 
     @Override
-    public double compareTo(Home another) {
-        return this.getArea() - another.getArea();
+    public int compareTo(Home another) {
+        return (int) (this.getArea() - another.getArea());
     }
 
     @Override
